+++ src/uwsgiconf/options/networking.py
@@ -135,6 +135,7 @@ class Networking(OptionsGroup):
     def set_socket_params(
             self,
             *,
+            timeout: int | None = None,
             send_timeout: int | None = None,
             keep_alive: bool | None = None,
             no_defer_accept: bool | None = None,
@@ -142,6 +143,8 @@ class Networking(OptionsGroup):
             buffer_receive: int | None = None
     ):
         """Sets common socket params.
+
+        :param timeout: Internal sockets timeout. Default: 4.
 
         :param send_timeout: Send (write) timeout in seconds.
 
@@ -157,6 +160,7 @@ class Networking(OptionsGroup):
         :param buffer_receive: Set SO_RCVBUF (bytes).
 
         """
+        self._set('socket-timeout', timeout)
         self._set('so-send-timeout', send_timeout)
         self._set('so-keepalive', keep_alive, cast=bool)
         self._set('no-defer-accept', no_defer_accept, cast=bool)